@@ -7,10 +7,8 @@
   "scripts": {
     "test": "jest",
     "lint": "eslint ./ --ext=\".js,.jsx\" --ignore-pattern build --ignore-pattern __tests__/ --ignore-pattern __mocks__/",
-    "webpack-server": "webpack-dev-server --hot --progress --colors",
-    "start": "npm run webpack-server",
-    "build-development": "webpack --optimize-minimize",
-    "build": "cross-env NODE_ENV=production webpack --optimize-minimize --optimize-dedupe"
+    "start": "webpack-dev-server",
+    "build": "cross-env NODE_ENV=production webpack --optimize-minimize"
   },
   "jest": {
     "moduleNameMapper": {
@@ -19,44 +17,50 @@
     }
   },
   "dependencies": {
-    "babel-polyfill": "^6.13.0",
-    "exports-loader": "^0.6.3",
-    "imports-loader": "^0.6.5",
+    "@blueprintjs/core": "^1.28.0",
+    "@blueprintjs/table": "^1.25.0",
+    "es6-promise": "^4.1.1",
+    "exports-loader": "^0.6.4",
+    "imports-loader": "^0.7.1",
     "material-ui": "^0.16.7",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "normalize.css": "^7.0.0",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-hot-loader": "next",
     "react-loader": "^2.2.0",
     "react-router": "^3.0.2",
     "react-tap-event-plugin": "^2.0.1",
-    "sussol-react-table": "^0.1.6"
+    "sussol-react-table": "^0.1.7",
+    "whatwg-fetch": "^0.11.0"
   },
   "devDependencies": {
     "babel-cli": "^6.22.2",
     "babel-core": "^6.6.5",
-    "babel-eslint": "^6.1.2",
+    "babel-eslint": "^8.0.0",
     "babel-jest": "^19.0.0",
-    "babel-loader": "^6.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-dynamic-import-node": "^1.1.0",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
     "cross-env": "^4.0.0",
-    "css-loader": "^0.23.1",
-    "enzyme": "^2.8.0",
-    "eslint": "^2.7.0",
-    "eslint-config-airbnb": "^7.0.0",
-    "eslint-plugin-jsx-a11y": "^0.6.2",
-    "eslint-plugin-react": "^4.3.0",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.28.0",
+    "css-loader": "^0.28.7",
+    "enzyme": "^2.9.1",
+    "eslint": "^4.7.1",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.3.0",
+    "file-loader": "^0.11.2",
+    "html-webpack-plugin": "^2.30.1",
     "jest": "^18.1.0",
-    "react-addons-test-utils": "^15.4.2",
-    "react-hot-loader": "^1.3.0",
+    "react-addons-test-utils": "^15.6.0",
     "react-test-renderer": "^15.5.4",
     "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.1",
-    "whatwg-fetch": "^0.11.0"
+    "style-loader": "^0.18.2",
+    "webpack": "^3.6.0",
+    "webpack-dev-server": "^2.8.2"
   },
   "author": "Sustainable Solutions (NZ)",
   "contributors": [
