@@ -1,7 +1,6 @@
 import React, { PureComponent } from 'react';
 
-import { ResultsTable } from './ResultsTable.jsx';
-import { SearchBar } from './SearchBar.jsx';
+import { SearchBar } from './SearchBar';
 import { ColumnFixedWidths } from '../settings';
 
 export class Search extends PureComponent {
@@ -15,8 +14,12 @@ export class Search extends PureComponent {
         key: 'code',
         title: 'Code',
       }],
+      loadedComponent: false,
+      resultsTableComponent: null,
       showResults: false,
     };
+
+    this.resultsTableComponent = null;
   }
 
   componentDidMount() {
@@ -40,6 +43,7 @@ export class Search extends PureComponent {
       return;
     }
     this.setState({ showResults: true, resultData: data, searchTerm });
+    this.renderLazyTable();
   }
 
   /**
@@ -59,8 +63,8 @@ export class Search extends PureComponent {
     const fixedWidthArray = Object.values(columnFixedWidths);
     // add total amount of fixed width to subtract from usable space
     const totalFixedWidth = fixedWidthArray.reduce((sum, width) => {
-      let sumClone = sum;
-      return (sumClone += width);
+      const newWidth = sum + width;
+      return newWidth;
     }, 0);
 
     // get useable lengths, minus any fixed table width data (columnFixedWidths)
@@ -79,7 +83,25 @@ export class Search extends PureComponent {
     return columnWidths;
   }
 
+  /**
+  * renderLazyTable
+  *
+  * Create a table component chunk and lazy load it into state.
+  * Blueprint.js table and its core are large files. Need to keep
+  * load times to a minimum.
+  * After all, users may not even search.
+  *
+  */
+  renderLazyTable() {
+    return import('./ResultsTable').then((table) => {
+      this.resultsTableComponent = table.ResultsTable;
+      this.setState({ loadedComponent: true });
+    });
+  }
+
   render(props) {
+    const ResultsTable = this.resultsTableComponent;
+
     return (
       <div className="search-area">
         <SearchBar
@@ -90,34 +112,37 @@ export class Search extends PureComponent {
           onSearchClear={() => this.handleSearchChange(null)}
         />
         {/* render results table */}
-        {this.state.showResults && <ResultsTable
-          {...props}
-          changeColumnWidths={(columns, tableWidth, columnFixedWidths) => (
-            this.handleCalculateColumnWidths(columns, tableWidth, columnFixedWidths)
-          )}
-          columnWidthSettings={ColumnFixedWidths}
-          columnWidths={this.state.columnWidths}
-          columns={this.state.columns}
-          data={this.state.resultData}
-          getTableHeight={(tableRef) => {
-            // covers cases of:
-            //   - mobile screens in portrait
-            //   - everything else that's _not_ mobile in landscape
-            let height = (document.documentElement.clientHeight - tableRef.offsetTop) - 100;
+        {
+          this.state.showResults && ResultsTable !== null &&
+          <ResultsTable
+            {...props}
+            changeColumnWidths={(columns, tableWidth, columnFixedWidths) => (
+              this.handleCalculateColumnWidths(columns, tableWidth, columnFixedWidths)
+            )}
+            columnWidthSettings={ColumnFixedWidths}
+            columnWidths={this.state.columnWidths}
+            columns={this.state.columns}
+            data={this.state.resultData}
+            getTableHeight={(tableRef) => {
+              // covers cases of:
+              //   - mobile screens in portrait
+              //   - everything else that's _not_ mobile in landscape
+              let height = (document.documentElement.clientHeight - tableRef.offsetTop) - 100;
 
-            // mobile screens in landscape only
-            const landscapeString = 'screen and (orientation: landscape) and (max-width: 820px)';
-            if (window.matchMedia(landscapeString).matches) {
-              height = tableRef.offsetTop;
-            }
+              // mobile screens in landscape only
+              const landscapeString = 'screen and (orientation: landscape) and (max-width: 820px)';
+              if (window.matchMedia(landscapeString).matches) {
+                height = tableRef.offsetTop;
+              }
 
-            // set the height
-            this.setState({ tableHeight: height });
-          }}
-          height={this.state.tableHeight}
-          numberOfResults={this.state.resultData.length}
-          searchTerm={this.state.searchTerm}
-        />}
+              // set the height
+              this.setState({ tableHeight: height });
+            }}
+            height={this.state.tableHeight}
+            numberOfResults={this.state.resultData.length}
+            searchTerm={this.state.searchTerm}
+          />
+        }
       </div>
     );
   }
