@@ -1,4 +1,4 @@
-import { About } from './About.jsx';
+import { About } from './About';
 
 export const AboutRoute = {
   path: '/about',
